Product deleted successfully

+++ CRUD_Coffee_Store/coffee-store-client/src/components/CoffeeCard/CoffeeCard.jsx
@@ -1,20 +1,54 @@
 import { useContext } from "react";
 import { CoffeeContext } from "../../Provider/ApiProvider";
+import Swal from "sweetalert2";
 
 /* eslint-disable react/prop-types */
 const CoffeeCard = () => {
   const { coffees, setCoffees } = useContext(CoffeeContext);
   const handleDelete = (id) => {
-    fetch(`http://localhost:5000/coffee/${id}`, {
-      method: "DELETE",
-    })
-      .then((response) => response.json())
-      .then((data) => {
-        console.log(data);
-        // kaj kora kano boglam na
-        const newCoffee = coffees.filter((coff) => coff._id !== id);
-        setCoffees(newCoffee);
-      });
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        fetch(`http://localhost:5000/coffee/${id}`, {
+          method: "DELETE",
+        })
+          .then((response) => response.json())
+          .then((data) => {
+            console.log(data);
+            if (data.deletedCount > 0) {
+              // kaj kora kano boglam na
+              const newCoffee = coffees.filter((coff) => coff._id !== id);
+              setCoffees(newCoffee);
+              Swal.fire({
+                title: "Deleted!",
+                text: "Your file has been deleted.",
+                icon: "success",
+              });
+            }
+          });
+      }
+    });
+
+    // fetch(`http://localhost:5000/coffee/${id}`, {
+    //   method: "DELETE",
+    // })
+    //   .then((response) => response.json())
+    //   .then((data) => {
+    //     console.log(data);
+    //     if (data.deletedCount > 0) {
+    //       const newCoffee = coffees.filter((coff) => coff._id !== id);
+    //       setCoffees(newCoffee);
+    //       alert("Deleted successfully");
+    //     }
+    //     // kaj kora kano boglam na
+    //   });
   };
   return (
     <div className="grid grid-cols-1 md:grid-cols-2 gap-8">
